+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- [Java, JavaScript] Add `Lineage.background()` and `.ruleBackground()` ([#140](https://github.com/cucumber/query/pull/140))
+- [Java, JavaScript] Add `Query.findAllTestCaseStartedOrderBy` and `.findAllTestCaseFinishedOrderBy` ([#114](https://github.com/cucumber/query/pull/114))
+### Removed
+- [JavaScript] BREAKING CHANGE: Remove defunct legacy methods from `Query` ([#141](https://github.com/cucumber/query/pull/141))
+
+## [14.7.0] - 2025-12-08
+### Added
+- [NET] Added .NET port  ([#81](https://github.com/cucumber/query/pull/81))
 
 ## [14.6.0] - 2025-10-27
 ### Added
@@ -269,7 +278,8 @@ from `@cucumber/messages` instead.
 ### Added
 - First JavaScript implementation
 
-[Unreleased]: https://github.com/cucumber/query/compare/v14.6.0...HEAD
+[Unreleased]: https://github.com/cucumber/query/compare/v14.7.0...HEAD
+[14.7.0]: https://github.com/cucumber/query/compare/v14.6.0...v14.7.0
 [14.6.0]: https://github.com/cucumber/query/compare/v14.5.0...v14.6.0
 [14.5.0]: https://github.com/cucumber/query/compare/v14.4.0...v14.5.0
 [14.4.0]: https://github.com/cucumber/query/compare/v14.3.0...v14.4.0